tests/wpt/web-platform-tests/html/webappapis/idle-callbacks/callback-invoked.html

<!doctype html>
<meta charset=utf-8>
<title>requestIdleCallback callback must be called eventually</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<div id="log"></div>
<script>
  async_test(function (t) {
    assert_false(document.hidden, "document.hidden must exist and be false to run this test properly");
    window.requestIdleCallback(t.step_func_done());
  }, "requestIdleCallback callback is invoked at least once before the timeout");
</script>
